backend/requirements.txt
alembic==1.14.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
appdirs==1.4.4
astor==0.8.1
asttokens==2.4.1
bcrypt==5.0.0
blinker==1.9.0
chinesecalendar==1.10.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
distlib==0.3.8
ecdsa==0.19.1
et-xmlfile==1.1.0
executing==2.0.1
fastapi==0.124.0
filelock==3.14.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
fonttools==4.53.0
fsspec==2024.6.1
greenlet==3.1.1
h11==0.16.0
holidays==0.63
idna==3.11
iniconfig==2.3.0
intel-openmp==2021.4.0
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
llvmlite==0.44.0
Mako==1.3.6
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mkl==2021.4.0
MouseInfo==0.1.3
mpmath==1.3.0
networkx==3.3
numba==0.61.0
numpy==1.26.4
openpyxl==3.1.2
packaging==24.1
pandas==2.2.2
pandasgui==0.2.15
parso==0.8.4
passlib==1.7.4
pbr==6.0.0
pillow==10.3.0
platformdirs==4.2.2
plotly==5.22.0
pluggy==1.6.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.11
pure-eval==0.2.2
pyarrow==16.1.0
pyasn1==0.6.1
PyAutoGUI==0.9.54
pydantic==2.12.5
pydantic_core==2.41.5
pygame==2.6.0
PyGetWindow==0.0.9
Pygments==2.18.0
PyMsgBox==1.0.9
pynput==1.7.7
pyparsing==3.1.2
pyperclip==1.9.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
PyRect==0.2.0
PyScreeze==0.1.30
pytest==9.0.2
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
pytweening==1.2.0
pytz==2024.1
pywin32==306
PyYAML==6.0.2
qtstylish==0.1.5
redis==5.0.1
rsa==4.9.1
setuptools==70.1.1
six==1.16.0
SQLAlchemy==2.0.36
sqlmodel==0.0.27
stack-data==0.6.3
starlette==0.50.0
stevedore==5.2.0
style==1.1.0
sympy==1.13.0
tbb==2021.13.0
tenacity==8.4.2
torch==2.3.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2024.1
update==0.0.1
uv==0.9.7
uvicorn==0.38.0
virtualenv==20.26.2
virtualenvwrapper-win==1.2.7
wcwidth==0.2.13
Werkzeug==3.1.3
wordcloud==1.9.3
xlrd==2.0.1
xlwings==0.31.7
